airflow/requirements.txt
kafka-python
fastapi
uvicorn
pandas==2.0.3
numpy==1.24.4
nltk
spacy==3.7.2
textblob
scikit-learn==1.3.2
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.2.2
beautifulsoup4
requests
aiohttp
Cython
setuptools
python-dotenv
